Add Hashtable
Added a hashtable node struct to construct an array of hash nodes.  This allows me to return the dataset name associated with the terminated child process, since I am not allowed to return anything but the sum and difference from a terminated process.

I use a linear search algorithm to find the correct key / value pair.

Also added calls to free memory once it is not needed.  This is to prevent a memory leak.