FCFS Naiive Implementation Complete
Completed my naiive implementation of the solution for FCFS.  Line 217 of FCFS was giving trouble, would increment the node pointer for the execute array, changing the value of that element to the one currently in the tempReady queue.  Fixed this by making the tempArr a pointer array instead, and passing the pointer instead of the value to it.  This now computes for the base dataset (CPULoad.dat) along with changes I made in sch-helpers.c