Fix Segfaults
Multiple segfaults occurred due to poor pointer management.  The tempArr was a local array to the while loop, which would get destroyed every time it enters the conditional it resides in.  This means any existing pointers to this data would be destroyed, causing us to reference null values.  Changed this so that: tempArr is now a pointer array, and is initialized at the top of the program.  This way pointers can be passed without being destroyed, as we don't reinstantiate the array.  This means I also needed to change qsort to pass the array to it to remove the segfault it invoked.  Also changed the top header include so it works on Linux (c for Windows) and changed the calculation value for MAX_CPUS to properly compute the CPU Utilization.