More FCFS Cleanup
Addressed an issue with the qsort function.  Changed the array of tempArr processes to be an array, instead of a pointer to an array.  This allows the qsort function to work properly.  I now compare the PIDs from the sorted tempArr and pass the actual process nodes from tempArray to the enqueue function to make sure the data moves accordingly.  This solved many issues, including not properly populating the finish array, an infinite loop, duplicate processes and more.  This version of the code works on the red server we need to test on and produces the closest output to the sample solution so far.