Write bytes in correct order
Was previously using a global offset to write the bytes in order of consume from the ring buffer, but realized the error and changed this to the actual offset value from the original text file.  This allowed the bytes to be placed in their respective order.

Also included the advanced solution, where we do not continue until all threads have completed without using a join to block.